services/ai/claims_automation/requirements.txt
# Dynamic Claims Automation Layer (DCAL) Dependencies
# Python 3.11+

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Async & Concurrency
asyncio==3.4.3
aiohttp==3.9.1
aiomysql==0.2.0
asyncpg==0.29.0

# Message Queue (Kafka)
aiokafka==0.10.0
kafka-python==2.0.2

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# ML & Data Science
scikit-learn==1.4.0
xgboost==2.0.3
lightgbm==4.2.0
numpy==1.26.3
pandas==2.1.4
scipy==1.11.4

# ML Explainability
shap==0.44.1

# Security & Crypto
cryptography==42.0.0
pyjwt[crypto]==2.8.0
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
passlib[bcrypt]==1.7.4

# HTTP & API
httpx==0.26.0
requests==2.31.0

# Monitoring & Observability
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0

# Logging
structlog==24.1.0
python-json-logger==2.0.7

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.26.0
faker==22.0.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
pyyaml==6.0.1
click==8.1.7

# Development
black==24.1.1
mypy==1.8.0
ruff==0.1.14


